servver/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
certifi
pyserial
scikit-learn
pandas
groq
aiohttp
gTTS
edge-tts
joblib
